Migrate label page from /docs/components/label to /components/label
Move Form and Disabled examples to the new reference page, delete the
old page, update routes/sidebar/home links, and rewrite E2E tests to
target the new URL.
